Toward strata() support within ppse.qr
Idea: insist that the terms to be swept out be named at the
beginning of the formula.
In the process I also swept out the intercept, in cases where no strata()
was given. This changes the answer even when strata() isn't involved,
increasing separation between ppse.glm and ppse.qr results, so that
tests are currently failing.  To be addressed...