Inline comment/note to self re sandwich calc
The bread calculation is being done in a somewhat odd manner;
it alarmed me when I looked at it after not having looked in a
while.  I don't know of a good reason to do it this way rather than
in the manner suggested in these new comments, which would more
closely mirror the formulas supporting the calculation; if I had
earlier found the alternate method to lead to instabilities, that
instability would probably have been addressed by now, in e1d5390.
Probably I should try out this change sooner or later.