$\hat{C}_n$-consistency calls for $p^2 = o(n^{-1})$, not just $p\log p = o(n^{-1})$
The tactics being used to control
$|\hat{A}_{n}(\hat\beta) - \hat{A}_{n}(\betatrue)|_2$
(in the proof of Lemma~\ref{lem:ChatC}) actually require $p^2/n = o(1)$,
not just $p(\log p)/n = o(1)$.
The same may be true for the similar tactic applied to
$|\hat{B}_{n}(\hat\beta) - \hat{B}_{n}(\betatrue)|_2$.
Previous versions maintaining otherwise
appear to have been mistaken.

The proof is adapted from Wang 2011, which makes still
stronger rate conditions, so perhaps this is not surprising.
Potentially a variation on the proof strategy could avoid
having to be backed into this corner.
For instance, asserting or exerting control over
$|\hat\beta - \betatrue|_1$, instead of or in addition
to $|\hat\beta - \betatrue|_2$, would do the trick here.