In light of $|x|_\infty$ condition, excise |X|_2 moment conditions
He and Shao's (2000) results don't put any universal bound on the
x-es, although they do require that that their squared and cubed
l2-norms be bounded. For other reasons it suits this project to impose
universal bounds on $|x|_\infty$, however, and He and Shao's weaker
conditions follow. (Boundedness of X's turns out to be difficult to do
without, specifically in Lemma~\ref{lem:ChatC}'s proof of
$|\hat{B}(\hat\beta) - \hat{B}(\betatrue)|_2 = o_P(1)$;
more centrally, in order for control of PISE's to translate to
differences in the underlying index score being forced to 0,
boundedness of the Xes is either necessary or necessary in light of
the KISS principle.)

This commit comments out {A-l2Sfinite}, which encoded a He-Shao type
condition as well as a separate condition now called {A-Sfullrank},
relacing references to it with references either to {A-Sfullrank} or
to {A-boundedXes} as appropriate. The new references to {A-boundedXes}
are accompanied by inline comments referencing {A-l2Sfinite}.